backend-drf/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Cache — Redis
redis==5.0.1
django-redis==5.4.0

# ML — version matched with model file
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.37
matplotlib==3.7.2
h5py==3.9.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Config
python-decouple==3.8